+++ tests/PdfExportFromViewDataTest.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Sfneal\ViewExport\Tests;
+
+use Dompdf\Exception;
+use Sfneal\ViewExport\Pdf\PdfExportService;
+use Sfneal\ViewExport\Tests\Traits\PdfExportValidations;
+
+class PdfExportFromViewDataTest extends TestCase
+{
+    use PdfExportValidations;
+
+    /**
+     * Setup the test environment.
+     *
+     * @return void
+     * @throws Exception
+     */
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->exporter = PdfExportService::fromViewData('test', ['string'=>"Here's a string!"]);
+    }
+}